# ColorPlugValueWidget : Lazy `ColorShooser` create