This improves the performance of creating a `ColorPlugValueWidget` by deferring the creation of the internal color chooser to when it is first made visible.

For a single `ColorPlugValueWidget`, the difference is not all that noticeable, but for a node such as `aiStandardSurface`, the improved creation time of all the widgets in the node editor is subtly faster.

It's also noticeably faster when testing on top of https://github.com/GafferHQ/gaffer/pull/6003 where we are adding a number of additional widgets and increasing the amount of time needed to populate the node editor when not deferring color chooser creation.

For me, some timings when selecting an `aiStandardSurface` node :
- Current code : 0.60 seconds
- Deferred creation : 0.35 seconds (41% shorter)
- New `ColorChooser` without deferred creation : 0.75 seconds (26% longer than current)
- New `ColorChooser` with deferred creation : 0.35 seconds (53% shorter than without deferred)

### Checklist ###

- [X] I have read the [contribution guidelines](https://github.com/GafferHQ/gaffer/blob/main/CONTRIBUTING.md).
- [X] I have updated the documentation, if applicable.
- [X] I have tested my change(s) in the test suite, and added new test cases where necessary.
- [X] My code follows the Gaffer project's prevailing coding style and conventions.